requirements.txt
altgraph==0.17.3
color-matcher==0.5.0
ddt==1.6.0
docutils==0.20.1
imageio==2.28.1
numpy==1.24.3
pefile==2023.2.7
Pillow==9.5.0
pyinstaller==5.11.0
pyinstaller-hooks-contrib==2023.3
pywin32-ctypes==0.2.0
